requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.1.0
python-socketio==5.9.0
requests==2.31.0
beautifulsoup4==4.12.0
pytest==7.4.0
pytest-django==4.5.2
black==23.3.0
flake8==6.0.0
